luma-orb/requirements.txt
pygame
requests
pyttsx3
SpeechRecognition
pyaudio
faster-whisper==0.10.0